requirements/dev.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.3
django-debug-toolbar==3.2.4
python-dotenv==0.19.2
sqlparse==0.4.2
tzdata==2021.5
